// src/Client/Logistics.Angular/projects/shared/src/lib/api/generated/models/tenant-feature.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export type TenantFeature = 'dashboard' | 'employees' | 'loads' | 'trucks' | 'customers' | 'invoices' | 'payments' | 'eld' | 'load_board' | 'messages' | 'notifications' | 'safety' | 'expenses' | 'payroll' | 'timesheets' | 'maintenance' | 'trips' | 'reports';